requirements.txt
flask==2.3.3
werkzeug==2.3.7
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.1
matplotlib==3.7.2
seaborn==0.12.2
pymongo==4.4.1
flask-wtf==1.1.1
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
Flask-PyMongo==2.3.0
flask-session==0.5.0
email-validator==2.0.0


